requirements.txt
Django==1.10.3
djangorestframework==3.5.3
Markdown==2.6.7
django-filter==1.0.0
psycopg2==2.6.2
